Fixed bit size issue for int64. Added test cases for typedefs and linked list

+++ playstate/playstate/types.h
@@ -23,8 +23,8 @@ namespace playstate
 	typedef int int32;
 	typedef unsigned int uint32;
 
-	typedef long int64;
-	typedef unsigned long uint64;
+	typedef long long int64;
+	typedef unsigned long long uint64;
 
 	typedef float float32;
 	typedef double float64;
